Clear D0 at the end of the AES trap
The existing code just cleared the low-order word; now all 32 bits
are cleared.  This circumvents a bug in WordPerfect 4.1 which
assumed that D0 was cleared after calling the AES.

+++ aes/gemsuper.c
@@ -51,7 +51,7 @@
 #define ENABLE_KDEBUG
 #endif
 
-extern WORD super(WORD cx, AESPB *pcrys_blk);   /* called only from gemdosif.S */
+extern LONG super(WORD cx, AESPB *pcrys_blk);   /* called only from gemdosif.S */
 
 GLOBAL WORD     gl_mnclick;
 
@@ -463,7 +463,7 @@ static void xif(AESPB *pcrys_blk)
  *  Supervisor entry point.  Stack frame must be exactly like
  *  this if supret is to work.
  */
-WORD super(WORD cx, AESPB *pcrys_blk)
+LONG super(WORD cx, AESPB *pcrys_blk)
 {
     switch(cx)
     {
